Made a new plot tool to do simple visualization of spikes and weights
