+++ spiking_network/plotting/__init__.py
@@ -1,0 +1,1 @@
+from spiking_network.plotting.plot import visualize_spikes, visualize_weights, visualize_time_dependence
