+++ spiking_network/plotting/plot_timings.py
