+++ spiking_network/plotting/plots/comp_with_rolling_for_memory.png
