+++ spiking_network/plotting/plots/cpu_vs_gpu.png
