+++ spiking_network/plotting/plots/mikkel_comp_with_and_without_parallelization.png
