+++ spiking_network/plotting/plots/old_vs_new_no_equi.png
