+++ spiking_network/plotting/plots/old_vs_new_sim.png
