+++ spiking_network/plotting/plots/parallelization.png
