+++ spiking_network/plotting/plots/parallelization_compare_across_neurons.png
